Add MainViewModel class
