+++ PlanetClock/PlanetClock/PlanetClock.csproj
@@ -73,6 +73,7 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="Animations.cs" />
+    <Compile Include="MainViewModel.cs" />
     <Compile Include="TextBlockFadeBehavior.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
